+++ CHANGELOG.md
@@ -3,6 +3,12 @@
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
+
+## 2.0.0-beta.4 - 2022-07-26
+### Fixed
+- business -> organization nomenclature change
+- Updated default Tax Adjuster  issue: #17
+
 ## 2.0.0-beta.3 - 2022-07-15
 ### Fixed
 - craft\element\Address issue: #16